Create RTU Esp32
Wi-Fi Connection: Connects the device to a specified Wi-Fi network using the SSID and password.

MQTT Client: Sets up an MQTT client that connects to a specified MQTT broker, and subscribes to a "flowRate" topic. When new messages are published on this topic, the payload is parsed and assigned to the receivedFlowRate variable.

Water Flow Rate Measurement: It uses a sensor to measure the water flow rate. The sensor generates pulses, and an interrupt service routine counts these pulses. This pulse count is used to calculate the water flow rate in milliliters per minute.

Leak Detection: It checks for water leaks based on the difference between the smoothed flow rate and the received flow rate. If the difference exceeds a certain threshold for a specified period, a leak is detected.

Leak Severity Check: It categorizes the leakage based on the difference in the flow rate, and it sends this information to the Blynk app.

Blynk App Integration: It uses the Blynk library to send the smoothed flow rate, leak detection status, and leak severity level to the Blynk app.

OTA Updates: It supports Over-The-Air (OTA) updates, enabling the firmware to be updated wirelessly.

Maintaining Wi-Fi Connection: It periodically checks the Wi-Fi connection and reconnects if necessary.

Maintaining MQTT Connection: If the device is not connected to the MQTT server, it attempts to reconnect.